Shorten some file and target names under reference/
This PR shortens `reference/InterpreterFoo.{cpp,h}` to just
`reference/Foo.{cpp,h}`, and applies a similar transformation to
target names in both CMake and Bazel builds.

The only name that's left untouched is `reference/InterpreterOps`
because shortening that would conflict with the already existing
`reference/Ops`.

There are some C++ names that could potentially be shortened as well,
e.g. InterpreterFallback and InterpreterConfiguration, but C++ names
are out of scope of this PR.