Fix VHLO DenseI64Array bug and refactor shape refinement pass (#1875)
VHLO bug: `DenseI64ArrayAttr` doesn't have a notion of splats like
`DenseIntElementsAttr` does. The following array when serialized as as
the VHLO wrapper around `DenseIntElementsAttr` could not be converted to
`DenseI64ArrayAttr` properly:

```
dense<0> : tensor<2xi64>
vs
array<i64: 0, 0>
```

There the value that gets serialized is just `0`. When converting to
`DenseI64ArrayAttr` must look at the type and see if the splat should be
expanded. This isn't a huge memory issue since these attrs are only used
in cases where the number of values in the array is small.

Shape refinement: Refactor into a header file with some patterns that
can be populated, instead of a single API entrypoint.