Port CHLO ops to I64DenseArrayOrElements1DAttr
This is an intermediate step in the migration to `DenseI64ArrayAttr`.

I had to move the definition for the attr to `Base.td` because it's now
used in `ChloOps.td` as well as `StablehloOps.td`.

I also had to do some shenanigans to support checking if the
`broadcast_dimensions` are legal in
`ReifyBroadcastBinaryOpReturnTypeShapes`. Because the attr is obtained
through generic means (`op->getAttr("broadcast_dimensions")`), it doesn't
go through the conversion from storage (e.g. when doing
`op.getBroadcastDimensions`, which returns `SmallVector<int64_t>`), so
both `DenseIntElementsAttr` and `DenseI64ArrayAttr` have to be handled.

Ideally, I would be able to do `dyn_cast<I64DenseArrayOrElements1DAttr>`
and use a custom method on the attr to get the integers, but I see no
way to define a custom method on an attr.

https://github.com/openxla/stablehlo/issues/1578