Remove location constraint on compact printing of ReduceOpt
During discussion of #1523 we came to the conclusion that it's
impossible to continously guarantee that location information is kept
forwrad and backwards through transformations.

There is already precendent in other transformations that location
information may be ellided.

Remove the constraint for the compact pretty printing of ReduceOpt that
requires the locations be the same. Update the tests as necessary.