Disallow reusing variables across CHECK-LABEL directives
Ran into this while integrating: internal configuration sets
`--enable-var-scope`, but we don't, so some tests failed internally but
not here. I think in general this is a good setting to use: it's easier
to understand a test if the only variables in scope are the variables
defined by that test.