Add support for layering_check and change default toolchain
Bazel has support for layering_check identification which helps validate
that sources are only pulling in header files from their direct
dependencies.

This is considered best practice in Bazel, since if you depend on
indirect dependencies it's easy to break upstream package consumers
unknowingly.

At the moment layering_check is only suported by Clang. The change also
changes the default CC variable to be Clang as a result.

TODO: We should consider adding a toolchain directly in Bazel to make it
hermetic or consider using Nix to bring in workspace tooling.