Move large categories of tests to their own folders (#1913)
The `tests` folder contains a lot of tests for various things. In
particular, there are a lot of tests for the interpreter and a lot of
tests related to VHLO, so IMO it makes sense to give these their own
folders. I find that this helps navigate the tests.

I considered creating folders for CHLO, printing related tests, and
verification related tests, but IMO there aren't enough of those to
really warrant it.