Use array literals where possible in MLIR source (#1930)
This change flips all the literals that can be written using `array`
instead of `dense` without requiring C++ changes. The purpose of this
change is to reduce the diff in a future change, where I will actually
change the op fields to use `DenseI64ArrayAttr` instead of
`I64DenseArrayOrElements1DAttr`. That will require making changes to
parsing and serialization, among others.