Fix `all_gather` to concat in order of process group (#1934)
Currently, the implementation concatenates rendezvous tensors which are
sorted by process ids. However, the spec implies that the order of
concatenation happens in the same order as the generated `processGroup`,
which is not necessarily sorted. This change reflects what is implied in
the spec.

Thanks @linuxlonelyeagle for filing the issue!

closes #1933