Avoid explicitly constructing compatible attributes in a few places (#1958)
Attrs will be implicitly copied. Avoid explicitly constructing the attrs
as this introduces a dependency on the type of the attrs, meaning the
code will need to be changed if the attrs' type changes. (Indeed I had
to change this code recently when we changed the attrs from
DenseIntElementsAttr to DenseI64ArrayAttr).