Use reference interpreter to eval ops in refine shapes pass
This change removes some duplication between the interpreter and the
refine shapes pass. I didn't go all the way (there is still some
duplication, e.g. the slice op is reimplemented in the refine shapes
pass), because I'm not sure if the purpose of the interpreter is _only_
to be used as a reference, or if we want to use it for passes too.

The implementation in the interpreter is of course more generic than
what is needed here, and we likely lose some efficiency due to the type
conversions. However, in my opinion it is more important in this case to
reduce duplication.