Add shellcheck lint to CI (#1971)
shellcheck is a shell script analysis tool that can find bugs and subtle
issues in shell scripts: https://www.shellcheck.net/

As part of adding the check, this commit also fixes any issues that are
currently surfaced by shellcheck. These fell into the following
categories:
- Adding missing shebangs to scripts because shellcheck recommendations
depend on the shell being used.
- [SC2086](https://www.shellcheck.net/wiki/SC2086) (info): Double quote
to prevent globbing and word splitting.
- [SC2034](https://www.shellcheck.net/wiki/SC2034) (warning): <variable>
appears unused. Verify use (or export if used externally).
- [SC2207](https://www.shellcheck.net/wiki/SC2207) (warning): Prefer
mapfile or read -a to split command output (or quote to avoid
splitting).
- [SC2185](https://www.shellcheck.net/wiki/SC2185) (info): Some finds
don't have a default path. Specify '.' explicitly.
- [SC2068](https://www.shellcheck.net/wiki/SC2068) (error): Double quote
array expansions to avoid re-splitting elements.
- [SC2145](https://www.shellcheck.net/wiki/SC2145) (error): Argument
mixes string and array. Use * or separate argument.
- [SC2236](https://www.shellcheck.net/wiki/SC2236) (style): Use -n
instead of ! -z.
- [SC2155](https://www.shellcheck.net/wiki/SC2155) (warning): Declare
and assign separately to avoid masking return values. (We have a few
cases of this where we define `readonly` variables using a command that
could fail.)

I am proposing that we use a lint level of `style`, which is the minimum
level provided by shellcheck (so we will be running all lints). Most of
the lints reported by shellcheck were of the `info` variety, due to
variables being expanded outside of quotes (e.g. `echo $x` instead of
`echo "$x"`). However there were a few warnings and errors as well, and
one "style" lint which I believe is a no-brainer (use `-n` instead of `!
-z`).

I am also adding suppression comments for the following:
- [SC2016](https://www.shellcheck.net/wiki/SC2016) (info): Expressions
don't expand in single quotes, use double quotes for that. (There are a
few places where we actually do not want `$...` to be expanded.)
- [SC1003](https://www.shellcheck.net/wiki/SC1003) (info): Want to
escape a single quote? echo 'This is how it'\''s done'. (This occurs in
a specific use of `sed` to add a newline to files: `sed -i -e '$a\'` --
apparently shellcheck doesn't know about this idiom.)