Fix SC2155 issues (#1975)
I forgot to include this in
https://github.com/openxla/stablehlo/pull/1971 (made the change locally
and forgot to push before merging).

I originally suppressed these lints because I misunderstood how `set -o
errexit` interacts with `readonly`. I thought a failure would still
bring down the whole script, but that's not the case. So instead of
suppressing the lint, I fixed the issue.