Enhance rendezvous docs and fix race condition (#1956)
The PR removes the use of `shared_ptr`s as it complicates the logic of
`rendezvous`. Instead of relying on the use count of `shared_ptr`, I've
added a variable to track the state separately instead. Now, there is no
need to worry about the scope of the `shared_ptr`s and cleaning up
resource before exiting.

Thanks @mlevesquedion for sharing code snippets offline to help simplify
this!