Perform initialization in pass initialize method (#1966)
This is unlikely to give us performance gains since most of our passes
run on modules anyway (so the initialization probably already occurs
only once), but it is cleaner to separate the initialization of a pass
from the actual running of the pass.

The code in `initialize` will run when the pass runs regardless of
whether there is at least one instance of the target operation. However,
modules are pretty much always present so this is unlikely to change
anything.