Add support for Bazel 6 & 7
Previous PR #2006 added support for Bazel 7 but some of the changes in
Bazel 7 are not backwards compatible with 6.

This is a problem for some of our consumers that embed our project in
other Bazel workspaces.

Specifically, the name of the runfiles directory was changed from the
workspace name to be '_main'.

The solution adopted here is to use @python_rules and programmatically
determine the runfiles directory relative to the lit.cfg.py file.