Add support for LLVM toolchain
We should strive to make our Bazel builds as hermetic as possible.
Bazel has support for toolchains
(https://bazel.build/extending/toolchains) which I admit are a bit more
advanced than I am familiar with.

Fortunately, there seems to be a community contribution, for a toolchain
that is provided for LLVM.

I picked a version that is quite old (11.1) since some of our developers
at Google are on 20.04 LTS.