Exclude MLIR bytecode files from license check
Also, don't add an extra line between each skip, and use the same
message for each type of skipped file (easier to read).

I tested the script locally by creating files with various suffixes
and confirmed that the skipped suffixes are indeed skipped (and files
with a non-skipped suffix are not skipped).