Add StableHLO dialect version for bytecode
This enables (at least) invalidating previously generated StableHLO bytecode
serialization (not using VHLO) by bumping the version. The version is
optionally set and if not set, not serialized - this would result in no changes
to what is serialized by default today unless explicitly set.

This would enable a better error message than the existing where a loaded
bytecode would fail and knowledge required as to why.

This does not represent a policy change nor an automatic test added.

Could also add to CHLO but wanted to discuss first.