Print clang-format version in CI (#2032)
The version of clang-format available in CI seems to be different from
the one on my machine, but I can't easily see what the version of
clang-format running in CI is, so I can't really debug. Print the
version so that moving forward if someone runs afoul of the check they
can repro and fix.