Make type error checks in tests less brittle
These error messages are automatically generated via ODS constraints so
we shouldn't test the entire message. In particular, I don't think it is
useful to test the part of the message that lists the possible element
types a given tensor might have. This exposes us to breakages if we e.g.
add a type that wasn't supported before.

Use regular expressions instead. Relevant documentation:
https://mlir.llvm.org/docs/Diagnostics/\#sourcemgr-diagnostic-verifier-handler