Clean up aggressive simplification and add tests (#2071)
I wanted to read this pass to familiarize myself with the code. In
passing, I made a few readability improvements here and there and while
making these I realized that we did not import the tests for this code.
Import the tests so we can have some confidence when making changes.

The tests are from
https://github.com/openxla/iree/blob/7782a414ea473c59f6d7a882cb510690ed666c79/compiler/plugins/input/StableHLO/stablehlo-iree/Conversion/Preprocessing/test/canonicalization.mlir,
with minor tweaks (formatting, reordering tests to match the
implementation, and adding missing test separators).