Fix stablehlo_legalize_to_vhlo.0_19_0.mlir bytecode
Also, use same header in all tests. Not sure why I used a different
header in the test for 0_19_0, I thought I copied it from 0_18_0, but it
seems I didn't. The current header doesn't check the bytecode, and it
turns out the bytecode is incorrect and/or incompatible with the
current implementation. The implementation works as desired, it is just
the bytecode that is messed up.