Fix broken links (#2082)
Most of these links are only broken when rendered out of tree, e.g. on
the OpenXLA website: https://openxla.org/stablehlo. For example, the
"More Examples" links lead to e.g.
https://openxla.org/stablehlo/tests/interpret/add.mlir, which is 404.

For links under `docs/` it is possible to use relative paths.

Also, add labels to images used in spec. This will materialize as the
"alt" text, which is used by screen readers.