Fix failed reify for ReverseOp (#2109)
The `ReverseOp` did not implement a `reifyReturnTypeShapes`. This
resulted in a failed assertion in the `getEmptyTensorFor` helper
function that gets used during the `stablehlo-legalize-to-linalg` pass
when the op has dynamic input/output. This patch implements this
function as well as adds a test for the newly supported case (input can
have dynamic dims as long as they don't correspond to the `dims`
attribute of the op)

Various other quality of life changes, such as having match failure
notifications have been added.