Implement ConditionallySpeculatable for all UnaryElementwise ops
A somewhat unfortunate thing is that we allow programs like:

```
func.func @foo(%arg: tensor<?xi64>) {
  %0 = stablehlo.abs %arg : (tensor<?xi64>) -> tensor<2xi64>
  return
}
```

Such a program is legal, but doesn't make much sense: in what scenario
would we know the output shape, but not the input shape? If it weren't
for such programs, the ops would all be `Pure`, but because at runtime
dynamic dimensions in the input type could differ from static dimensions
in the output type, these ops have to be `ConditionallySpeculatable`.