Support sparsity through StableHLO => Linalg lowering (#2149)
After StableHLO => Linalg lowering was implemented in the [linalg
RFC](https://github.com/openxla/stablehlo/pull/1610), most of the ops
stated by @aartbik in the [sparsity
RFC](https://github.com/openxla/stablehlo/pull/1143) can generate
correct linalg code with sparse encodings. And it can then be lowered
through mlir-opt with --sparsifier pipeline and run with mlir-runner.

The `sparse.mlir` file in this PR checks the generated linalg code with
sparse encodings. These are the same tests for the dense ops but with
sparse input/output.

The only op that couldn't be lowered to correct linalg with sparsity is
`stablehlo.concatenate`. Therefore, after discussing with @GleasonK, I
added a flag `enable-sparse-ops` to the linalg pass and rewrite
`stablehlo.concatenate` to `sparse_tensor.concatenate` when the flag is
set to true and input/output have sparse encodings. If there're more
linalg converters that need modification, a new pass for sparsity will
be created to run before linalg pass.