Remove Pure trait from Cholesky
The spec says "If there exists i where the input matrix is
not an Hermitian positive-definite matrix, then the behavior is
undefined." We could try to see if the inputs are constant and then
try to determine if they are Hermitian positive-define, and only then
report that the op is speculatable, but that seems like a heavy check
for a situation that isn't likely to occur frequently, so IMO it is
more reasonable to say that this op is not speculatable.

Pure is "NoMemoryEffect + AlwaysSpeculatable". If an op doesn't
implement ConditionallySpeculatable, then it is never Speculatable, so we
replace Pure with NoMemoryEffect, because these ops don't have memory
effects.