Augment bazelrc to support asan & ubsan
A prior failure in the google3 internal codebase showed that we were not
executing the sanitizers like we thought we were.

This change augments bazelrc to do asan and ubsan via config.
It leverages the default feature set offered by bazelrc.

I tested the commit priot to 474caa8370dccc753e4693a2ff4221864b227597 to
validate that the bug was exhibited.