Explain how constraints interact with dynamism (#2210)
This is for (P3) of the dynamism RFC:
https://github.com/openxla/stablehlo/blob/main/rfcs/20230704-dynamism-101.md#p3

The RFC says to update the "Notation" section of the spec, but I think
it makes more sense to have a separate section for dynamism, which I
plan to expand in future PRs. For now, I am adding some explanations
about constraint verification and undefined behavior.

(P3) also requires auditing existing verifiers and shape functions to
make sure that we allow dynamic dimension sizes everywhere. I am doing
this as part of writing tests for (P2).