Implement ConditionallySpeculatable for Fft
The logic is the same as for unary ops, except that if the type of FFT
is RFFT or IRFFT then the last `fft_length.size()` dimensions in the
operand need to be static, because there are extra constraints on them
that could turn out to be false at runtime if they are dynamic.

Indeed, there is a general constraint that `shape(result) = shape(operand)`. For
RFFT, the operand type element type is float, so the last `fft_length.size()` dims
in the operand have to be static, because `fft_length` is static (it is
an attribute). For IRFFT, the result element type is float, so the last
`fft_length.size()` dims in the result are inferred from the
`fft_length`, and they need to match the operand, so those dims have to
be static in the operand. There are also constraints on the last dimension,
but `size(fft_length) >= 1` so that is already covered by the previous check.