Add a note at type_inference.md for adding custom verifier if return type inference is not feasible (#2158)
if return type of the op is inferable, it can be created using
`op::build` which does not take result type as input. For such ops it is
recommended to move verification logic to shape functions to avoid
missing on any verification during op creation. If return type inference
is not feasible use custom verifier.