Add script to generate pass documentation
Ideally, we would have some fully automated process to generate the
documentation, which would run as a GitHub action on submit. However,
since we currently store all our docs in the repo itself (and they are
mirrored on the OpenXLA website by some external process), I think it
makes sense to use a manually-run script to generate the docs and check
that the docs are up to date in CI. Indeed, we rarely add new passes or
update existing pass documentation, so this shouldn't be too onerous.