Remove TraceOp from StableHLO (#2295)
Part of #2283. This op doesn't seem to have any uses in frameworks or
compilers. It shouldn't cause any large issues to remove this op all
together, also since this op was never specced, it is exempt from
compatibility guarantees.

In the case that it is used somewhere, I would recommend migration to a
custom_call.