[stablehlo] Add batching dims to `stablehlo.gather` and `stablehlo.scatter` (#2259)
Add `operand_batching_dims` and `start_indices_batching_dims` attributes
to `stablehlo.gather`. `operand_batching_dims` refers to the dimensions
of the `operand` that are treated as batch.
`start_indices_batching_dims` refers to the dimensions of the
`start_indices` that are treated as batch. The corresponding dimension
sizes must be equal. The semantics is equivalent to concatenating the
outputs of the gather with each slices of `operand` and `start_indices`.

Similarly, add `input_batching_dims` and `scatter_indices_batching_dims`
attributes to `stablehlo.scatter`. `input_batching_dims` refers to the
dimensions of each tensor in `inputs` that are treated as batch.
`scatter_indices_batching_dims` refers to the dimensions of the
`scatter_indices` that are treated as batch.

See https://github.com/openxla/stablehlo/pull/2084 for more information

---------

Co-authored-by: Kevin Gleason <gleasonk@google.com>
Co-authored-by: Gunhyun Park <gunhyun@google.com>
Co-authored-by: Gunhyun Park <ghpvnist28@gmail.com>
Co-authored-by: Abhinav Gunjal <agunjal@google.com>
Co-authored-by: mlevesquedion <mlevesquedion@google.com>