Port mhlo-quant-legalize-to-int pass to StableHLO (#2383)
This is the port of
[mhlo-quant-legalize-to-int](https://github.com/openxla/xla/blob/9449b0851c855ae3f17e8675d95d1f1b68104abd/xla/mlir_hlo/mhlo/transforms/mhlo_passes.td#L378)
pass to Stablehlo.



A bunch of tests needs to be updates based on:
- Different error message because of the differences in quantization
verification for mhlo and stablehlo.
- Update check directive to accommodate pretty printing in stablehlo vs
not pretty printed ops in mhlo (like dot_general)
- Modify tests with boradcast_in_dim to update the syntax of
broadcast_dimension (from `DenseIntElementsAttr` to `I64ArrayAttr`)
- Convert a test with per-axis stablehlo.max with per-tensor
stablehlo.max as max only supports per-tensor. The intent of the tests
is preserved.

Also, updated the original code to address creation of attributes of
type `I64ArrayAttr` instead of `DenseIntElementsAttr`.