Align canonicalization of stablehlo.broadcast_in_dim with XLA (#2371)
This patch is a two-stage:
1) change `isIotaRange` constraint to `llvm::sorted` in order to enable
the rewrite from `broadcast_in_dim` -> `reshape`:
```
  %0 = stablehlo.broadcast_in_dim %arg0, dims = [1, 2] : (tensor<3x6xi32>) -> tensor<1x3x6xi32>
```
is transformed into:
```
  %0 = stablehlo.reshape %arg0 : (tensor<3x6xi32>) -> tensor<1x3x6xi32>
```
(hlo_ops.cpp: BroadcastInDimSimplifier uses same approach)

2) `broadcast_in_dim` operator can be quite complex in its semantics,
sometimes functioning as a combination of `transpose` and `broadcast`.
In order to simplify analysis and optimization, it can be beneficial to
fully expand operand dimensions by ensuring that the size of
broadcast_dimensions matches the rank of the result. This can be
achieved by reshaping the input operand while preserving the constraint
`size(broadcast_dimensions) = rank(operand)` that the size of
broadcast_dimensions matches the rank of the operand.

```
    %3 = stablehlo.broadcast_in_dim %arg3, dims = [2, 1, 4] : (tensor<32x64x16xf32>) -> tensor<4x64x32x2x16xf32>
```
is transformed into:
```
  %2 = stablehlo.reshape %arg3 : (tensor<32x64x16xf32>) -> tensor<1x32x64x1x16xf32>
  %3 = stablehlo.broadcast_in_dim %2, dims = [0, 2, 1, 3, 4] : (tensor<1x32x64x1x16xf32>) -> tensor<4x64x32x2x16xf32>
```

Later it can be simplified even more by extracting transpose logic into
standalone operation, something like this:
```
  %2 = stablehlo.reshape %arg3 : (tensor<32x64x16xf32>) -> tensor<1x32x64x1x16xf32>
  %3 = stablehlo.transpose %2, dims = [0, 2, 1, 3, 4] : (tensor<1x32x64x1x16xf32>) -> tensor<1x64x32x1x16xf32>
  %4 = stablehlo.broadcast_in_dim %3, dims = [0, 1, 2, 3, 4] : (tensor<1x64x32x1x16xf32>) -> tensor<4x64x32x2x16xf32> 
```
Since `broadcast` is on the way out from stablehlo.

UPD: only patch 1 is accepted.