Python bindings for registering check dialect (#2445)
Check dialect is an auxiliary dialect used in StableHLO repository for
validation of StableHLO program evaluation. Currently there is no
cleaner way to parse a module containing check dialect operations or to
create them. One way to go around this is to textually modify the check
dialect ops to normalize them to generic MLIR text and allow
unregistered dialect to pars that text. However, this PR provodes a
cleaner way to process check dialect.

This PR prepares for open sourcing some of the utilities, leveraged in
https://github.com/openxla/stablehlo/pull/2404, for auto-generating
testdata formatted test files.