Remove type-inference dependency while creating qdq pattarns (#2460)
[ParentPR](https://github.com/openxla/stablehlo/pull/2459)

Previously while creating the QDQ pattern we use `create` API without
using the result type and hence reply on the type inference to derive
the result type. That works for most of the element-wise operations,
however, for dot_general and convolution the result type can might be
infeasible to infer in the presense of input quantize types.

The PR fixes that.

Note to the reviewers: To may just focus on the very last commit of the
chain. The rest is coming from parent PR.

[ChildPR](https://github.com/openxla/stablehlo/pull/2461)