Support qdq decomposition of DotGeneralOp and ConvolutionOp (#2461)
[ParentPR](https://github.com/openxla/stablehlo/pull/2460)

The PR 
1. adds patterns for qdq decomposition patterns for `DotGeneralOp` and
`ConvolutionOp`.
1. Updates the tests in
stablehlo/tests/stablehlo_legalize_quant_to_int.mlir updating only
negatives tests, which are previously unhanded by
stablehlo-legalize-quant-to-int. The current
stablehlo-legalize-quant-to-math uses the fallback to handle these
cases.

Note to the reviewers: To may just focus on the very last commit of the
chain. The rest is coming from parent PR.

[childPR](https://github.com/openxla/stablehlo/pull/2462)